Pull optimization profiles and expression equivalence disbatchers out of deriver.rs
